web_code/requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
python-dotenv==0.20.0

